+++ template/helper/Form.php
@@ -192,7 +192,7 @@ class Form extends \lithium\template\Helper {
 		parent::_init();
 
 		if ($this->_context) {
-			$this->_context->handlers(['wrap' => '_attributes']);
+			$this->_context->handlers(['wrap' => 'attributes']);
 		}
 	}